requirements.txt
pre-commit==3.4.0
pymongo==4.5.0
python-dotenv==1.0.0
python-telegram-bot==20.6
ruff==0.0.292
